requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlmodel>=0.0.24
sqlalchemy>=2.0.0,<2.1.0
pydantic>=2.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
python-multipart
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
jinja2>=3.1.0
itsdangerous>=2.1.0
